Pair internal/admin test files with split source files
