+++ internal/admin/playeractions.go
@@ -1,22 +1,12 @@
 package admin
 
 import (
-	"context"
 	"errors"
-	"fmt"
 	"log/slog"
 	"net/http"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 
 	"github.com/starquake/topbanana/internal/auth"
-	"github.com/starquake/topbanana/internal/bgtasks"
-	"github.com/starquake/topbanana/internal/csrf"
 	"github.com/starquake/topbanana/internal/handlers"
-	"github.com/starquake/topbanana/internal/mailer"
-	"github.com/starquake/topbanana/internal/render"
 )
 
 // HandlePlayerMarkVerified handles POST /admin/players/{playerID}/verify.
@@ -59,588 +49,6 @@ func HandlePlayerMarkVerified(
 		writeAudit(r.Context(), logger, store, actor.ID, playerID, auth.AdminActionVerify, nil)
 		flash.SetNotice(w, "Player marked verified.")
 		redirectToPlayerDetail(w, r, playerID)
-	})
-}
-
-// HandlePlayerResendVerification handles
-// POST /admin/players/{playerID}/resend-verification. Per-target
-// rate-limited (one resend per minute per playerID) so a stuck operator
-// hitting the button does not turn the admin page into a mail floodgate.
-// The send itself runs on a detached goroutine so the response is not
-// held open while SMTP dials.
-func HandlePlayerResendVerification(
-	logger *slog.Logger,
-	store auth.AdminPlayerStore,
-	tokens auth.VerifyTokenStore,
-	sender auth.VerifyEmailSender,
-	baseURL string,
-	limiter *PerTargetLimiter,
-	flash *auth.SignedFlash,
-	tasks *bgtasks.Tracker,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		playerID, ok := handlers.ParseIDFromPath(w, r, logger, "playerID")
-		if !ok {
-			return
-		}
-		actor, ok := requireAdminActor(w, r)
-		if !ok {
-			return
-		}
-
-		detail, ok := loadActionTarget(w, r, logger, store, playerID)
-		if !ok {
-			return
-		}
-		if detail.OnboardingState != auth.OnboardingStateUnverified || detail.Email == "" {
-			flash.SetError(w, "Resend is only available for unverified players with an email on file.", 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		wait, allowed, token := limiter.Allow(playerID)
-		if !allowed {
-			seconds := max(int((wait+time.Second-1)/time.Second), 1)
-			flash.SetError(w, "Slow down: wait a moment before resending.", seconds)
-			w.Header().Set("Retry-After", strconv.Itoa(seconds))
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		if !dispatchAdminResendVerification(
-			r.Context(), logger, tokens, sender, baseURL, detail.Email, playerID, tasks,
-		) {
-			// No mail went out (email not configured), so roll the stamp
-			// back: an operator on a misconfigured instance is not throttled
-			// for a send that never happened (#996). Token-matched so a
-			// second concurrent caller's newer stamp is not clobbered.
-			limiter.Cancel(playerID, token)
-			flash.SetError(w, "Email sending is not configured; no verification email was sent.", 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-		writeAudit(r.Context(), logger, store, actor.ID, playerID, auth.AdminActionResendVerification, nil)
-		flash.SetNotice(w, "Verification email dispatched.")
-		redirectToPlayerDetail(w, r, playerID)
-	})
-}
-
-// HandlePlayerSetEmail handles POST /admin/players/{playerID}/email.
-// Validates with auth.LooksLikeEmail, rejects collisions with the
-// existing ErrEmailTaken sentinel, and clears email_verified_at so the
-// changed address starts unverified - the operator then marks-verified
-// separately or triggers a resend (per #450 spec).
-func HandlePlayerSetEmail(
-	logger *slog.Logger,
-	store auth.AdminPlayerStore,
-	flash *auth.SignedFlash,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		playerID, ok := handlers.ParseIDFromPath(w, r, logger, "playerID")
-		if !ok {
-			return
-		}
-		actor, ok := requireAdminActor(w, r)
-		if !ok {
-			return
-		}
-		if !parseActionForm(w, r, logger, "set-email") {
-			return
-		}
-
-		email := strings.ToLower(strings.TrimSpace(r.PostFormValue("email")))
-		if !auth.LooksLikeEmail(email) {
-			flash.SetError(w, "Enter a valid email address.", 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		err := store.SetPlayerEmail(r.Context(), playerID, email)
-		switch {
-		case err == nil:
-			writeAudit(r.Context(), logger, store, actor.ID, playerID,
-				auth.AdminActionEmailSet, map[string]string{"new_email": email})
-			flash.SetNotice(w, "Email updated. The player still needs to verify the new address.")
-		case errors.Is(err, auth.ErrEmailTaken):
-			flash.SetError(w, "Another account already uses that email.", 0)
-		case errors.Is(err, auth.ErrPlayerNotFound):
-			flash.SetError(w, "Player not found.", 0)
-		default:
-			logger.ErrorContext(r.Context(), "error setting player email", slog.Any("err", err))
-			flash.SetError(w, "Could not update email. Try again.", 0)
-		}
-		redirectToPlayerDetail(w, r, playerID)
-	})
-}
-
-// HandlePlayerSetDisplayName handles POST /admin/players/{playerID}/display-name.
-// Admin only (gated by RequireAdmin at the route). Renames the target row to
-// the supplied display name; the store enforces the empty / taken sentinels so
-// the handler only maps them to flashes.
-func HandlePlayerSetDisplayName(
-	logger *slog.Logger,
-	store auth.AdminPlayerStore,
-	flash *auth.SignedFlash,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		playerID, ok := handlers.ParseIDFromPath(w, r, logger, "playerID")
-		if !ok {
-			return
-		}
-		actor, ok := requireAdminActor(w, r)
-		if !ok {
-			return
-		}
-		if !parseActionForm(w, r, logger, "set-display-name") {
-			return
-		}
-
-		name := strings.TrimSpace(r.PostFormValue("display_name"))
-
-		_, err := store.AdminRenamePlayer(r.Context(), playerID, name)
-		switch {
-		case err == nil:
-			writeAudit(r.Context(), logger, store, actor.ID, playerID,
-				auth.AdminActionDisplayNameSet, map[string]string{"new_displayName": name})
-			flash.SetNotice(w, "Display name updated.")
-		case errors.Is(err, auth.ErrDisplayNameEmpty):
-			flash.SetError(w, "Enter a display name.", 0)
-		case errors.Is(err, auth.ErrDisplayNameTaken):
-			flash.SetError(w, "That display name is already taken.", 0)
-		case errors.Is(err, auth.ErrPlayerNotFound):
-			flash.SetError(w, "Player not found.", 0)
-		default:
-			logger.ErrorContext(r.Context(), "error setting player displayName", slog.Any("err", err))
-			flash.SetError(w, "Could not update display name. Try again.", 0)
-		}
-		redirectToPlayerDetail(w, r, playerID)
-	})
-}
-
-// HandlePlayerSetPassword handles POST /admin/players/{playerID}/password.
-// Admin only (gated by RequireAdmin at the route). Rotates the target's
-// password and bumps session_version, signing the target out of their other
-// sessions. The raw password is never logged or echoed.
-func HandlePlayerSetPassword(
-	logger *slog.Logger,
-	store auth.AdminPlayerStore,
-	flash *auth.SignedFlash,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		playerID, ok := handlers.ParseIDFromPath(w, r, logger, "playerID")
-		if !ok {
-			return
-		}
-		actor, ok := requireAdminActor(w, r)
-		if !ok {
-			return
-		}
-		if !parseActionForm(w, r, logger, "set-password") {
-			return
-		}
-
-		password := r.PostFormValue("password")
-		if len(password) < auth.MinPasswordLength {
-			flash.SetError(w, fmt.Sprintf("Password must be at least %d characters.", auth.MinPasswordLength), 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-		if len(password) > auth.MaxPasswordLength {
-			flash.SetError(w, fmt.Sprintf("Password must be at most %d characters.", auth.MaxPasswordLength), 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		hash, err := auth.HashPassword(password)
-		if err != nil {
-			logger.ErrorContext(r.Context(), "error hashing player password", slog.Any("err", err))
-			flash.SetError(w, "Could not set password. Try again.", 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		err = store.ChangePlayerPassword(r.Context(), playerID, hash)
-		switch {
-		case err == nil:
-			writeAudit(r.Context(), logger, store, actor.ID, playerID, auth.AdminActionPasswordSet, nil)
-			flash.SetNotice(w,
-				"Password set. The player's other sessions were signed out; hand the new password over out-of-band.")
-		case errors.Is(err, auth.ErrPlayerNotFound):
-			flash.SetError(w, "Player not found.", 0)
-		default:
-			logger.ErrorContext(r.Context(), "error setting player password", slog.Any("err", err))
-			flash.SetError(w, "Could not set password. Try again.", 0)
-		}
-		redirectToPlayerDetail(w, r, playerID)
-	})
-}
-
-// roleIsValid reports whether the "role" form value is one of the three
-// accepted tiers (#538).
-func roleIsValid(role string) bool {
-	switch role {
-	case auth.RolePlayer, auth.RoleHost, auth.RoleAdmin:
-		return true
-	default:
-		return false
-	}
-}
-
-// HandlePlayerSetRole handles POST /admin/players/{playerID}/role (#538).
-// Admin only (gated by RequireAdmin at the route). Reads the desired tier from
-// the "role" form field, diffs it against the target's current role, and
-// applies the change. The last-admin guard refuses a change that would remove
-// the only remaining Admin. One admin_audit row (role_changed) is written per
-// real change, with the payload carrying {from, to}.
-func HandlePlayerSetRole(
-	logger *slog.Logger,
-	store auth.AdminPlayerStore,
-	sender auth.VerifyEmailSender,
-	mailConfigured bool,
-	flash *auth.SignedFlash,
-	tasks *bgtasks.Tracker,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		playerID, ok := handlers.ParseIDFromPath(w, r, logger, "playerID")
-		if !ok {
-			return
-		}
-		actor, ok := requireAdminActor(w, r)
-		if !ok {
-			return
-		}
-		if !parseActionForm(w, r, logger, "set-role") {
-			return
-		}
-
-		desired := r.PostFormValue("role")
-		if !roleIsValid(desired) {
-			flash.SetError(w, "Choose a role of player, host, or admin.", 0)
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		detail, ok := loadActionTarget(w, r, logger, store, playerID)
-		if !ok {
-			return
-		}
-
-		from := detail.Role
-		if from == desired {
-			flash.SetNotice(w, "Role unchanged.")
-			redirectToPlayerDetail(w, r, playerID)
-
-			return
-		}
-
-		removingAdmin := from == auth.RoleAdmin && desired != auth.RoleAdmin
-		if removingAdmin {
-			if !writeAdminDemotion(w, r, logger, store, flash, playerID, desired) {
-				return
-			}
-		} else if !writeRoleChange(w, r, logger, store, flash, playerID, desired) {
-			return
-		}
-
-		writeAudit(r.Context(), logger, store, actor.ID, playerID, auth.AdminActionRoleChanged,
-			map[string]string{"from": from, "to": desired})
-
-		notice := roleChangeNotice(desired) +
-			maybeNotifyRoleChange(r, logger, sender, mailConfigured, detail, desired, playerID, tasks)
-		flash.SetNotice(w, notice)
-		redirectToPlayerDetail(w, r, playerID)
-	})
-}
-
-// writeAdminDemotion persists a demotion away from Admin via the atomic
-// store guard: a missing target is flashed as "not found", the last-admin
-// refusal as "promote another first", any other store error as a flashed
-// retry. The count-and-update happen in one statement in the store, so two
-// concurrent demotions of the two remaining admins cannot both pass (#997).
-// Returns true only when the demotion was written.
-func writeAdminDemotion(
-	w http.ResponseWriter, r *http.Request, logger *slog.Logger,
-	store auth.AdminPlayerStore, flash *auth.SignedFlash, playerID int64, desired string,
-) bool {
-	err := store.DemoteAdmin(r.Context(), playerID, desired)
-	switch {
-	case err == nil:
-		return true
-	case errors.Is(err, auth.ErrLastAdmin):
-		flash.SetError(w, "Cannot remove the last admin - promote another first.", 0)
-		redirectToPlayerDetail(w, r, playerID)
-	case errors.Is(err, auth.ErrPlayerNotFound):
-		flash.SetError(w, "Player not found.", 0)
-		redirectToPlayerDetail(w, r, playerID)
-	default:
-		logger.ErrorContext(r.Context(), "error demoting admin", slog.Any("err", err))
-		flash.SetError(w, "Could not update role. Try again.", 0)
-		redirectToPlayerDetail(w, r, playerID)
-	}
-
-	return false
-}
-
-// writeRoleChange persists the role and maps a write failure onto its
-// response: a missing target is a 404, any other store error a flashed
-// retry + 303. Returns true only when the role was written.
-func writeRoleChange(
-	w http.ResponseWriter, r *http.Request, logger *slog.Logger,
-	store auth.AdminPlayerStore, flash *auth.SignedFlash, playerID int64, desired string,
-) bool {
-	err := store.SetPlayerRole(r.Context(), playerID, desired)
-	switch {
-	case err == nil:
-		return true
-	case errors.Is(err, auth.ErrPlayerNotFound):
-		// Match the other player-action handlers: flash + 303 back, not a bare 404.
-		flash.SetError(w, "Player not found.", 0)
-		redirectToPlayerDetail(w, r, playerID)
-	default:
-		logger.ErrorContext(r.Context(), "error setting player role", slog.Any("err", err))
-		flash.SetError(w, "Could not update role. Try again.", 0)
-		redirectToPlayerDetail(w, r, playerID)
-	}
-
-	return false
-}
-
-// roleLabel maps a role constant to its human word, so the success
-// flash and the notification email name the tier the same way.
-func roleLabel(role string) string {
-	switch role {
-	case auth.RoleAdmin:
-		return "admin"
-	case auth.RoleHost:
-		return "host"
-	default:
-		return "player"
-	}
-}
-
-// roleChangeNotice is the success flash naming the new tier.
-func roleChangeNotice(role string) string {
-	return "Player role set to " + roleLabel(role) + "."
-}
-
-// maybeNotifyRoleChange honours the opt-in checkbox: when it is unset it
-// returns an empty string so the flash is unchanged. When set, it
-// dispatches the role-change notice (if SMTP is configured and the target
-// has a verified email) and returns the sentence to append to the success
-// flash. A player without a verified address, or an instance without
-// email configured, gets no mail and a flash that says so.
-//
-//nolint:revive // mailConfigured reflects whether SMTP is wired (an instance fact carried from mailer.StatusView), not a behavioural toggle the caller flips per request.
-func maybeNotifyRoleChange(
-	r *http.Request,
-	logger *slog.Logger,
-	sender auth.VerifyEmailSender,
-	mailConfigured bool,
-	detail *auth.PlayerDetail,
-	desired string,
-	playerID int64,
-	tasks *bgtasks.Tracker,
-) string {
-	if r.PostFormValue("notify_email") == "" {
-		return ""
-	}
-	if !mailConfigured {
-		return " Email is not configured, so no notification was sent."
-	}
-	if detail.Email == "" || detail.EmailVerifiedAt == nil {
-		return " The player has no verified email, so no notification was sent."
-	}
-	dispatchRoleChangeNotice(r.Context(), logger, sender, detail.Email, desired, playerID, tasks)
-
-	return " A notification email was sent to the player."
-}
-
-// dispatchRoleChangeNotice sends a best-effort notice to the player that
-// an administrator changed their role. Mirrors notifyOldAddressOfChange:
-// the send runs on a detached goroutine with a bounded timeout so a
-// closed browser tab does not cancel it and SMTP latency is not
-// observable from the redirect; a failure is logged at Warn and never
-// blocks the response.
-func dispatchRoleChangeNotice(
-	ctx context.Context,
-	logger *slog.Logger,
-	sender auth.VerifyEmailSender,
-	recipient, role string,
-	playerID int64,
-	tasks *bgtasks.Tracker,
-) {
-	msg := mailer.Message{
-		To:      recipient,
-		Subject: "Your Top Banana! account role changed",
-		Body: "An administrator changed the role on your Top Banana! account to " + roleLabel(role) + ".\n\n" +
-			"If you have questions about this change, contact the site administrator.\n",
-		Kind: mailer.KindRoleChangeNotice,
-	}
-	bg, cancel := context.WithTimeout(context.WithoutCancel(ctx), mailer.SendTimeout+15*time.Second)
-	tasks.Go(func() {
-		defer cancel()
-		if err := sender.Send(bg, msg); err != nil {
-			logger.WarnContext(bg, "role change notice dispatch failed",
-				slog.Int64("player_id", playerID), slog.Any("err", err))
-		}
-	})
-}
-
-// playerCreatePageData backs the playernew.gohtml template.
-type playerCreatePageData struct {
-	Title       string
-	DisplayName string
-	Email       string
-	Error       string
-}
-
-// HandlePlayerCreateForm renders GET /admin/players/new.
-func HandlePlayerCreateForm(logger *slog.Logger, csrfMgr *csrf.Manager) http.Handler {
-	renderer := NewTemplateRenderer(logger, csrfMgr, "admin/pages/playernew.gohtml")
-
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		renderer.Render(w, r, http.StatusOK, playerCreatePageData{
-			Title: "Admin Dashboard - New Player",
-		})
-	})
-}
-
-// HandlePlayerCreateSubmit handles POST /admin/players. Creates the row
-// with email_verified_at stamped so the new account can log in
-// immediately; the admin hands the credentials over out-of-band. Empty
-// password is rejected because a player who cannot log in is not what
-// this action is for.
-func HandlePlayerCreateSubmit(
-	logger *slog.Logger,
-	csrfMgr *csrf.Manager,
-	store auth.AdminPlayerStore,
-	flash *auth.SignedFlash,
-) http.Handler {
-	renderer := NewTemplateRenderer(logger, csrfMgr, "admin/pages/playernew.gohtml")
-
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		actor, ok := requireAdminActor(w, r)
-		if !ok {
-			return
-		}
-		r.Body = http.MaxBytesReader(w, r.Body, maxFormSize)
-		if err := r.ParseForm(); err != nil {
-			logger.InfoContext(r.Context(), "create-player form parse failed", slog.Any("err", err))
-			renderer.Render(w, r, http.StatusBadRequest, playerCreatePageData{
-				Title: "Admin Dashboard - New Player",
-				Error: "Form was malformed or too large.",
-			})
-
-			return
-		}
-
-		input := newPlayerInput(r)
-		if input.errMsg != "" {
-			renderer.Render(w, r, http.StatusBadRequest, playerCreatePageData{
-				Title:       "Admin Dashboard - New Player",
-				DisplayName: input.DisplayName, Email: input.Email,
-				Error: input.errMsg,
-			})
-
-			return
-		}
-
-		hash, err := auth.HashPassword(input.Password)
-		if err != nil {
-			logger.ErrorContext(r.Context(), "error hashing password for admin-create", slog.Any("err", err))
-			renderer.Render(w, r, http.StatusInternalServerError, playerCreatePageData{
-				Title:       "Admin Dashboard - New Player",
-				DisplayName: input.DisplayName, Email: input.Email,
-				Error: "Could not create player. Try again.",
-			})
-
-			return
-		}
-
-		player, err := store.CreatePlayerByAdmin(r.Context(), input.DisplayName, input.Email, hash)
-		if err != nil {
-			renderCreatePlayerError(w, r, renderer, input, err)
-
-			return
-		}
-
-		writeAudit(r.Context(), logger, store, actor.ID, player.ID, auth.AdminActionCreated,
-			map[string]string{"new_email": input.Email})
-		flash.SetNotice(w, fmt.Sprintf(
-			"Player %q created with email %s. Hand the password over out-of-band; it is not stored here.",
-			player.DisplayName, input.Email,
-		))
-		redirectToPlayerDetail(w, r, player.ID)
-	})
-}
-
-// newPlayerCreateInput is the trimmed + validated form data for the
-// admin-create-player flow. errMsg carries the first validation
-// failure; the handler re-renders the form with that banner.
-type newPlayerCreateInput struct {
-	DisplayName string
-	Email       string
-	Password    string
-	errMsg      string
-}
-
-func newPlayerInput(r *http.Request) newPlayerCreateInput {
-	in := newPlayerCreateInput{
-		DisplayName: strings.TrimSpace(r.PostFormValue("display_name")),
-		Email:       strings.ToLower(strings.TrimSpace(r.PostFormValue("email"))),
-		Password:    r.PostFormValue("password"),
-	}
-	if in.DisplayName == "" {
-		in.DisplayName = auth.GeneratePetname()
-	}
-	if !auth.LooksLikeEmail(in.Email) {
-		in.errMsg = "Enter a valid email address."
-
-		return in
-	}
-	if len(in.Password) < auth.MinPasswordLength {
-		in.errMsg = fmt.Sprintf("Password must be at least %d characters.", auth.MinPasswordLength)
-
-		return in
-	}
-	if len(in.Password) > auth.MaxPasswordLength {
-		in.errMsg = fmt.Sprintf("Password must be at most %d characters.", auth.MaxPasswordLength)
-	}
-
-	return in
-}
-
-// renderCreatePlayerError maps the store-level conflict sentinels onto
-// the form's re-render path. Anything else is a 500 from the
-// renderer's perspective; the caller logs the underlying err
-// separately for ops.
-func renderCreatePlayerError(
-	w http.ResponseWriter, r *http.Request, renderer *render.Renderer, in newPlayerCreateInput, err error,
-) {
-	status := http.StatusInternalServerError
-	msg := "Could not create player. Try again."
-	switch {
-	case errors.Is(err, auth.ErrDisplayNameTaken):
-		status = http.StatusConflict
-		msg = "That display name is already taken."
-	case errors.Is(err, auth.ErrEmailTaken):
-		status = http.StatusConflict
-		msg = "Another account already uses that email."
-	default:
-		// Fall through to the generic 500 message; err is logged by the caller.
-	}
-	renderer.Render(w, r, status, playerCreatePageData{
-		Title:       "Admin Dashboard - New Player",
-		DisplayName: in.DisplayName, Email: in.Email,
-		Error: msg,
 	})
 }
 
@@ -698,106 +106,4 @@ func loadActionTarget(
 	}
 
 	return detail, true
-}
-
-// dispatchAdminResendVerification mints + dispatches the verification
-// email on a detached goroutine. Mirrors auth.dispatchVerifyEmail so a
-// closed browser tab does not cancel the send and SMTP latency is not
-// observable from the redirect timing. Returns false without dispatching
-// when email is not configured (nil tokens/sender or empty baseURL) so
-// the caller can skip the audit row and the success notice.
-func dispatchAdminResendVerification(
-	ctx context.Context,
-	logger *slog.Logger,
-	tokens auth.VerifyTokenStore,
-	sender auth.VerifyEmailSender,
-	baseURL, recipient string,
-	playerID int64,
-	tasks *bgtasks.Tracker,
-) bool {
-	if tokens == nil || sender == nil {
-		return false
-	}
-	if baseURL == "" {
-		logger.WarnContext(ctx, "admin resend verification skipped: BASE_URL is empty",
-			slog.Int64("player_id", playerID))
-
-		return false
-	}
-	bg, cancel := context.WithTimeout(context.WithoutCancel(ctx), mailer.SendTimeout+15*time.Second)
-	tasks.Go(func() {
-		defer cancel()
-		auth.SendVerifyEmailBestEffort(bg, logger, tokens, sender,
-			baseURL, recipient, playerID, time.Now().UTC())
-	})
-
-	return true
-}
-
-// PerTargetLimiter is a per-player-id cool-down for admin actions that
-// dispatch outbound mail. Concurrency-safe; the map is pruned of stale
-// entries every Allow call so memory stays proportional to the live
-// caller set. Same shape as auth.VerifyResendLimiter but keyed on
-// playerID instead of source IP.
-type PerTargetLimiter struct {
-	mu     sync.Mutex
-	last   map[int64]time.Time
-	window time.Duration
-	now    func() time.Time
-}
-
-// NewPerTargetLimiter returns a limiter using the supplied window and
-// [time.Now] as the clock. The clock is injectable via the export_test
-// seam so tests can fast-forward without sleeping.
-func NewPerTargetLimiter(window time.Duration) *PerTargetLimiter {
-	return newPerTargetLimiterWithClock(window, time.Now)
-}
-
-func newPerTargetLimiterWithClock(window time.Duration, now func() time.Time) *PerTargetLimiter {
-	return &PerTargetLimiter{
-		last:   map[int64]time.Time{},
-		window: window,
-		now:    now,
-	}
-}
-
-// Allow reports whether target may dispatch right now. On admit stamps
-// the bucket so the next call within the window is blocked and returns
-// the stamp as a token; pass it to [PerTargetLimiter.Cancel] to revert
-// this specific stamp on a downstream bail-out. On block returns the
-// remaining wait so the caller can render it, and the zero-time token.
-func (l *PerTargetLimiter) Allow(target int64) (time.Duration, bool, time.Time) {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-
-	now := l.now()
-	cutoff := now.Add(-2 * l.window)
-	for k, ts := range l.last {
-		if ts.Before(cutoff) {
-			delete(l.last, k)
-		}
-	}
-	if prev, ok := l.last[target]; ok {
-		if remaining := l.window - now.Sub(prev); remaining > 0 {
-			return remaining, false, time.Time{}
-		}
-	}
-	l.last[target] = now
-
-	return 0, true, now
-}
-
-// Cancel reverts the stamp Allow wrote for target, but only if the live
-// stamp still matches token. Matching on token keeps a slow caller from
-// clobbering a newer stamp a second concurrent caller wrote in between
-// this caller's Allow and Cancel - that newer stamp must stand so the
-// second caller's window is honoured. Idempotent: Cancel on a target
-// with no entry, or with a stale token, is a no-op.
-func (l *PerTargetLimiter) Cancel(target int64, token time.Time) {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-
-	if cur, ok := l.last[target]; ok && cur.Equal(token) {
-		delete(l.last, target)
-	}
 }
